refactor compents make navigation reuseable create route for landingpage
