+++ client/src/App.js
@@ -3,19 +3,20 @@ import { Route, Switch, Redirect } from "react-router-dom";
 import { GlobalStyles } from "./components/index";
 import DashBoard from "./components/views/DashBoard";
 import LoginPage from "./components/views/LoginPage";
-import LandingPage from "./components/landing-page/LandingPage";
+import LandingPage from "./components/LandingPage/LandingPage";
+import Navigation from "./components/Navigation/Navigation";
 
 function App() {
   return (
     <div>
       <GlobalStyles />
-      <LandingPage />
+      <Navigation />
       <Switch>
+        <Route exact path="/" component={LandingPage} />
         <Route exact path="/dashboard" component={DashBoard} />
         <Route exact path="/login" component={LoginPage} />
         <Redirect to="/login" />
       </Switch>
-      
     </div>
   );
 }
