+++ client/src/components/LandingPage/LandingPage.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+//components
+// import Navigation from "../navigation/Navigation";
+import HeroSection from "./HeroSection/HeroSection";
+
+const LandingPage = () => {
+  return (
+    <div className="main-container">
+      {/* <Navigation /> */}
+      <HeroSection/>
+    </div>
+  );
+};
+
+export default LandingPage;
