+++ client/src/components/Navigation/NavigationStyles.js
@@ -1,4 +1,4 @@
-import { createGlobalStyle } from "styled-components";
+import { createGlobalStyle } from 'styled-components';
 
 export const NavigationStyles = createGlobalStyle`
   header {
