--- client/src/components/landing-page/LandingPage.js
@@ -1,16 +1,0 @@
-import React from "react";
-
-//components
-import Navigation from "./navigation/Navigation";
-import HeroSection from "./hero-section/HeroSection";
-
-const LandingPage = () => {
-  return (
-    <div className="main-container">
-      <Navigation />
-      <HeroSection/>
-    </div>
-  );
-};
-
-export default LandingPage;
